Moved suggestion.js to new renderer/lib directory, updated tests
Auditor: @bbondy
